tests/wpt/tests/css/css-text/parsing/text-spacing-trim-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing text-spacing-trim with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-text-4/#text-spacing-trim-property">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("text-spacing-trim", "normal");
test_invalid_value("text-spacing-trim", "normal space-all");
test_invalid_value("text-spacing-trim", "normal space-first");
test_invalid_value("text-spacing-trim", "auto space-all");
test_invalid_value("text-spacing-trim", "auto space-first");
</script>
</body>
</html>
